Pathfinding
I think the right files have been added/uploaded.

Added my Pathfinding algorithm, and created a new Scene for now which
shows how my Algorithm works. Also moved the relevant scripts
appropriate "Task 1" folder.

Still have Comments to add, and to give the overall thing a much nicer
look. Also need to work on it to allow it use to use the data that is
parsed in from the map, instead of using objects.